+++ moim/src/main/java/com/youth/moim/domain/user/Gender.java
@@ -1,4 +1,4 @@
-package com.youth.moim.domain;
+package com.youth.moim.domain.user;
 
 public enum Gender {
     FEMALE, MALE
