+++ moim/src/main/java/com/youth/moim/domain/user/UserStore.java
@@ -1,4 +1,4 @@
-package com.youth.moim.domain;
+package com.youth.moim.domain.user;
 
 public interface UserStore {
     void registerUser(User entity);
