+++ moim/src/main/java/com/youth/moim/infrastructure/user/UserStoreImpl.java
@@ -1,7 +1,7 @@
-package com.youth.moim.infrastructure;
+package com.youth.moim.infrastructure.user;
 
-import com.youth.moim.domain.User;
-import com.youth.moim.domain.UserStore;
+import com.youth.moim.domain.user.User;
+import com.youth.moim.domain.user.UserStore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
