+++ moim/src/main/java/com/youth/moim/presentation/user/UserRequest.java
@@ -1,8 +1,8 @@
 package com.youth.moim.presentation.user;
 
-import com.youth.moim.domain.Gender;
-import com.youth.moim.domain.MoimRule;
-import com.youth.moim.domain.User;
+import com.youth.moim.domain.user.Gender;
+import com.youth.moim.domain.user.MoimRule;
+import com.youth.moim.domain.user.User;
 
 public class UserRequest {
     public record SignInOrganizer(
